Components > ImagesSearchForm: form submission now handled by prop so parent components can use form for multiple purposes
So this can be used in ImagesPage as well as on the
BlogImagesList.
